moved view mode menu to TimelineMenu
