+++ src/Components/TimelineMenu/TimelineMenu.js
@@ -5,7 +5,12 @@ import MenuIcons from '../Icons/MenuIcons';
 import DownloadExcel from '../Export/ToExcel';
 import CreateNewEvent from "../NewEvent/NewEventComponent";
 import PermissionsModal from "../permissionsModal/permissionsModal";
-import {enableEditAction, disableEditAction} from "../../Actions/siteActions";
+import {
+    enableEditAction,
+    disableEditAction,
+    storyModeTimelineAction,
+    storyModeTableAction, storyModePrevTableAction
+} from "../../Actions/siteActions";
 import {
     controlNewEventModalAction,
     controlPermissionsModalAction,
@@ -19,6 +24,8 @@ import StoryInSearch from './Search/StorySearch';
 import UploadXlsxModal from './UploadXlsxModal/UploadXlsxModal';
 import TagsModal from '../Tags/TagsModal';
 import TagsSearch from "./Search/TagsSearch";
+import {getQueryStringParams} from "../../Actions/queryStringActions";
+import {withRouter} from "react-router";
 
 const { SubMenu } = Menu;
 
@@ -27,130 +34,173 @@ class TimelineMenu extends React.Component {
         super(props);
         const darkCheck = (this.props.darkMode === "true") || (this.props.darkMode === true);
         this.state = {
-            current: 'mail',
             menuTheme: darkCheck ? "dark" : "light",
             visiblePop: false,
 
         };
     };
 
-  handleClick = e => {
-    // console.log('click ', e);
-    this.setState({
-      current: e.key,
-    });
-  };
-
-
-  cancel = e => {
-  console.log(e);
-  message.error('Click on No');
+    setUrlParam = (key, value) => {
+        const pathName = this.props.history.location.pathname;
+        let currentSearchQuery = getQueryStringParams(this.props.history.location.search);
+        currentSearchQuery[key] = value;
+        this.props.history.push(
+            {
+                pathname: pathName,
+                search: "?" + new URLSearchParams(
+                    {...currentSearchQuery}
+                ).toString()
+            });
     };
 
-  confirm = e => {
-  console.log(e);
-  message.success('Click on Yes');
+    handleTimelineMode = () => {
+        this.props.storyModeTimelineAction();
+        this.setUrlParam('view', 'timeline');
+    };
+
+    handleTableMode = () => {
+        this.props.storyModeTableAction();
+        this.setUrlParam('view', 'full_table');
+    };
+
+    handlePreviewTableMode = () => {
+        this.props.storyModePrevTableAction();
+        this.setUrlParam('view', 'preview_table');
+    };
+
+
+    cancel = e => {
+        console.log(e);
+        message.error('Click on No');
+    };
+
+    confirm = e => {
+        console.log(e);
+        message.success('Click on Yes');
     };
 
     render() {
         const menuTheme = this.props.DarkMode === true ? "dark": "light";
         // console.log('Dark theme timeline', this.props.DarkMode);
-    return (
-        <div>
-      <Menu
-          onClick={this.handleClick}
-          mode="horizontal"
-          align="center"
-          selectable={false}
-          theme={menuTheme}
-      >
-          {(["write", "owner", "creator"].indexOf(this.props.role) === -1) ? null :
-          <Menu.Item key={"m_add"} icon={MenuIcons["plus"]} onClick={() => this.props.showNewEventModal()}>
-                    Add Event
-         </Menu.Item>}
+        return (
+            <div>
+                <Menu
+                    mode="horizontal"
+                    align="center"
+                    selectable={false}
+                    theme={menuTheme}
+                >
+                    {(["write", "owner", "creator"].indexOf(this.props.role) === -1) ? null :
+                        <Menu.Item key={"m_add"} icon={MenuIcons["plus"]} onClick={() => this.props.showNewEventModal()}>
+                            Add Event
+                        </Menu.Item>}
 
-        <SubMenu icon={MenuIcons["filter"]} title="Filter">
-            <Menu.Item disabled key="filter_by_time" >
+                    <SubMenu icon={MenuIcons["filter"]} title="Filter">
+                        <Menu.Item disabled key="filter_by_time" >
 
-            <StoryRangePicker />
-            </Menu.Item>
-            <Menu.Item disabled key="filter_by_word">
-                <StoryInSearch/>
-            </Menu.Item>
-            <Menu.Item disabled key="filter_by_tag">
-                <TagsSearch/>
-            </Menu.Item>
-        </SubMenu>
-          {(["write", "owner", "creator"].indexOf(this.props.role) === -1) ? null :
-              <SubMenu key="m_edit" icon={MenuIcons['edit']} title="Edit">
-                  {!this.props.editMode?
-                  <Menu.Item key="m_enable_edit" onClick={() => this.props.enableEdit()}>
-                      Enable Edit
-                  </Menu.Item>:
-                  <Menu.Item key="m_disable_edit" onClick={() => this.props.disableEdit()}>
-                      Disable Edit
-                  </Menu.Item>
-                  }
-                  <Menu.Item key="m_tags" onClick={() => this.props.showTagsModal()}>
-                      Tags
-                  </Menu.Item>
-                  <Menu.Item key="m_del_timeline" style={{color:"red"}} onClick={() => this.props.showDeleteTimelineModal()}>
+                            <StoryRangePicker />
+                        </Menu.Item>
+                        <Menu.Item disabled key="filter_by_word">
+                            <StoryInSearch/>
+                        </Menu.Item>
+                        <Menu.Item disabled key="filter_by_tag">
+                            <TagsSearch/>
+                        </Menu.Item>
+                    </SubMenu>
 
-                      Delete Story
+                    <SubMenu key="view" icon={MenuIcons['eye']} title="View">
 
-                  </Menu.Item>
-              </SubMenu>
-          }
+                        <Menu.Item
+                            key="timeline_mode"
+                            icon={MenuIcons['nodeindex']}
+                            onClick={() => this.handleTimelineMode()}>
+                            Timeline
+                        </Menu.Item>
+                        <Menu.Item
+                            key="prev_table_mode"
+                            icon={MenuIcons['compress']}
+                            onClick={() => this.handlePreviewTableMode()}>
+                            Preview Table
+                        </Menu.Item>
 
-         <SubMenu key="m_more" icon={MenuIcons['setting']} title="More">
-             {(["owner", "creator"].indexOf(this.props.role) === -1)? null:
-          <Menu.Item key={"m_permissions"} icon={MenuIcons["user"]}
-                             onClick={() => this.props.showPermissionsModal()}>
-                Permissions
-         </Menu.Item>}
+                        <Menu.Item
+                            key="table_mode"
+                            icon={MenuIcons['table']}
+                            onClick={() => this.handleTableMode()}>
+                            Full Table
+                        </Menu.Item>
 
-         <SubMenu key="export" icon={MenuIcons['download']} title="Export">
-              <Menu.Item key="export_excel" icon={MenuIcons['excel']}
-                         onClick={() => DownloadExcel(this.props.url, this.props.jwtToken)}>
-                      Excel
-            </Menu.Item>
+                    </SubMenu>
+                    {(["write", "owner", "creator"].indexOf(this.props.role) === -1) ? null :
+                        <SubMenu key="m_edit" icon={MenuIcons['edit']} title="Edit">
+                            {!this.props.editMode?
+                                <Menu.Item key="m_enable_edit" onClick={() => this.props.enableEdit()}>
+                                    Enable Edit
+                                </Menu.Item>:
+                                <Menu.Item key="m_disable_edit" onClick={() => this.props.disableEdit()}>
+                                    Disable Edit
+                                </Menu.Item>
+                            }
+                            <Menu.Item key="m_tags" onClick={() => this.props.showTagsModal()}>
+                                Tags
+                            </Menu.Item>
+                            <Menu.Item key="m_del_timeline" style={{color:"red"}} onClick={() => this.props.showDeleteTimelineModal()}>
 
-            </SubMenu>
+                                Delete Story
 
-          {(["write", "owner", "creator"].indexOf(this.props.role) === -1)? null:
-          <Menu.Item key={"m_upload"} icon={MenuIcons["upload"]}
-                             onClick={() => this.props.showUploadXlsxModal()}>
-                Import
-         </Menu.Item>}
-         </SubMenu>
+                            </Menu.Item>
+                        </SubMenu>
+                    }
 
-      </Menu>
-            <CreateNewEvent url={this.props.url} />
-            <PermissionsModal url={this.props.url}/>
-            <DeleteTimelineModal
-                url={this.props.url}
-                timelineId={this.props.timelineId}/>
-            <UploadXlsxModal
-                urlAddress={this.props.url}
-                timelineId={this.props.timelineId}/>
-            <TagsModal
-                url={this.props.url}
-                timelineId={this.props.timelineId}
-                       role={this.props.role}
-            />
+                    <SubMenu key="m_more" icon={MenuIcons['setting']} title="More">
+                        {(["owner", "creator"].indexOf(this.props.role) === -1)? null:
+                            <Menu.Item key={"m_permissions"} icon={MenuIcons["user"]}
+                                       onClick={() => this.props.showPermissionsModal()}>
+                                Permissions
+                            </Menu.Item>}
+
+                        <SubMenu key="export" icon={MenuIcons['download']} title="Export">
+                            <Menu.Item key="export_excel" icon={MenuIcons['excel']}
+                                       onClick={() => DownloadExcel(this.props.url, this.props.jwtToken)}>
+                                Excel
+                            </Menu.Item>
+
+                        </SubMenu>
+
+                        {(["write", "owner", "creator"].indexOf(this.props.role) === -1)? null:
+                            <Menu.Item key={"m_upload"} icon={MenuIcons["upload"]}
+                                       onClick={() => this.props.showUploadXlsxModal()}>
+                                Import
+                            </Menu.Item>}
+                    </SubMenu>
+
+                </Menu>
+                <CreateNewEvent url={this.props.url} />
+                <PermissionsModal url={this.props.url}/>
+                <DeleteTimelineModal
+                    url={this.props.url}
+                    timelineId={this.props.timelineId}/>
+                <UploadXlsxModal
+                    urlAddress={this.props.url}
+                    timelineId={this.props.timelineId}/>
+                <TagsModal
+                    url={this.props.url}
+                    timelineId={this.props.timelineId}
+                    role={this.props.role}
+                />
             </div>
 
 
-    );
-  }
+        );
+    }
 }
 const mapStateToProps = state => {
-  return {
-      DarkMode: state.sitesReducer.DarkMode,
-      editMode: state.sitesReducer.editMode,
-      jwtToken: state.usersReducer.jwtToken,
+    return {
+        DarkMode: state.sitesReducer.DarkMode,
+        editMode: state.sitesReducer.editMode,
+        jwtToken: state.usersReducer.jwtToken,
 
-  }
+    }
 };
 
 const mapDispatchToProps = dispatch => {
@@ -162,8 +212,11 @@ const mapDispatchToProps = dispatch => {
         showDeleteTimelineModal: () => {dispatch(controlDeleteTimelineModalAction(true))},
         showUploadXlsxModal: () => {dispatch(controlUploadXlsxModalAction(true))},
         showTagsModal: () => {dispatch(controlTagsModalAction(true))},
+        storyModeTimelineAction: () => {dispatch(storyModeTimelineAction())},
+        storyModeTableAction: () => {dispatch(storyModeTableAction())},
+        storyModePrevTableAction: () => {dispatch(storyModePrevTableAction())},
 
     }
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(TimelineMenu);
+export default connect(mapStateToProps, mapDispatchToProps)(withRouter(TimelineMenu));
